+++ xpocket-runtime/src/main/java/com/perfma/xlab/xpocket/command/impl/AttachCommand.java
@@ -1,5 +1,6 @@
 package com.perfma.xlab.xpocket.command.impl;
 
+import com.perfma.xlab.xpocket.command.AbstractSystemCommand;
 import com.perfma.xlab.xpocket.spi.command.CommandInfo;
 import com.perfma.xlab.xpocket.spi.context.SessionContext;
 import com.perfma.xlab.xpocket.spi.process.XPocketProcess;
