+++ xpocket-runtime/src/main/java/com/perfma/xlab/xpocket/command/impl/CdCommand.java
@@ -1,5 +1,6 @@
 package com.perfma.xlab.xpocket.command.impl;
 
+import com.perfma.xlab.xpocket.command.AbstractSystemCommand;
 import com.perfma.xlab.xpocket.framework.spi.impl.XPocketStatusContext;
 import com.perfma.xlab.xpocket.plugin.manager.PluginManager;
 import com.perfma.xlab.xpocket.spi.command.CommandInfo;
