+++ xpocket-runtime/src/main/java/com/perfma/xlab/xpocket/command/impl/HistoryCommand.java
@@ -1,5 +1,6 @@
 package com.perfma.xlab.xpocket.command.impl;
 
+import com.perfma.xlab.xpocket.command.AbstractSystemCommand;
 import com.perfma.xlab.xpocket.spi.command.CommandInfo;
 import com.perfma.xlab.xpocket.spi.process.XPocketProcess;
 import org.jline.reader.History;
