+++ xpocket-runtime/src/main/java/com/perfma/xlab/xpocket/framework/spi/impl/agent/AgentPluginLoaderImpl.java
@@ -18,7 +18,6 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FilenameFilter;
 import java.io.InputStreamReader;
-import java.io.PrintStream;
 import java.lang.instrument.Instrumentation;
 import java.net.URL;
 import java.nio.charset.Charset;
